[1738] Move search to `dashboard` application
