+++ amy/dashboard/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import path, include
 
 from dashboard import views
 
@@ -6,10 +6,15 @@ urlpatterns = [
     path('', views.dispatch, name='dispatch'),
 
     # admin dashboard main page
-    path('admin/', views.admin_dashboard, name='admin-dashboard'),
+    path('admin/', include([
+        path('', views.admin_dashboard, name='admin-dashboard'),
+        path('search/', views.search, name='search'),
+    ])),
 
     # trainee dashboard and trainee-available views
-    path('trainee/', views.trainee_dashboard, name='trainee-dashboard'),
-    path('trainee/training_progress/', views.training_progress, name='training-progress'),
-    path('trainee/autoupdate_profile/', views.autoupdate_profile, name='autoupdate_profile'),
+    path('trainee/', include([
+        path('', views.trainee_dashboard, name='trainee-dashboard'),
+        path('training_progress/', views.training_progress, name='training-progress'),
+        path('autoupdate_profile/', views.autoupdate_profile, name='autoupdate_profile'),
+    ])),
 ]
